Crate node_browser_testing

Running

Running this test can be done with

wasm-pack test --firefox --release --headless bin/node/browser-testing

or (without wasm-pack)

CARGO_TARGET_WASM32_UNKNOWN_UNKNOWN_RUNNER=wasm-bindgen-test-runner WASM_BINDGEN_TEST_TIMEOUT=60 cargo test --target wasm32-unknown-unknown

For debug infomation, such as the informant, run without the --headless flag and open a browser to the url that wasm-pack test outputs. For more infomation see https://rustwasm.github.io/docs/wasm-pack/.

Functions

__wbgt_runs_0